Derive unbundled Chooser from system version.
This is a shortcut to bootstrap the unbundling work by
providing an architectural equivalent to the next "phase"
of unbundling while we continue investigating (i.e., in
in this version, the unbundled ChooserActivity is
responsible for showing the Sharesheet UI, and then it
launches the user's selection via the startActivityAsCaller
API).

For more information about:

 * The "unbundling" project: go/sharesheet-unbundling

 * The architecture of this next (third) unbundling phase:
   go/sharesheet-unbundling-phases

 * The motivation for starting with this temporary
   inheritance-based implementation: b/202165481

Test: Manually tested (in combination with system-side
change ag/16285975). I had to disable the auto-selection
logic to see any UI (b/206831012), but I was able to get a
sharesheet activity from userspace, where I selected a
target that then got launched. This is *not* at parity,
but it's a baseline we can use when we investigate
the remaining issues.

Bug: 202165481
Change-Id: Ia2b89cc0e2735c8be7ad4cfc46c626a93e7fdee3