New Chooser report AppPrediction availability
The unbundled ChooserActivity can't query the App Prediction
Service availability directly, but as of ag/16306725 the
"bundled" (system) implementation performs the query and
includes the result in the handoff to the unbundled version.
This CL overrides the query method so that the unbundled
implementation instead just reports the value of this bit.

Test: Manually confirmed service now found to be available
Change-Id: I2a5bf8e2e8c644ca82eaf7b81061ec7de2d523b6