Remove support for "phase 2 delegation."
I had some idea about trying to support both interfaces to
the delegate chooser by inspecting the Intent, but I think
that's a bad idea (since we may mis-classify Intents that
should still be routed to the "full UI" chooser); it's
unnecessary; and it's a poor fit for our upcoming "phase 4."

A subsequent CL will remove the system-side dispatch. It's
safe to do that separately since the system dispatch should
never trigger anyways; were the conditions in place to
enable the "phase 2" delegation, the system chooser would
instead switch over to the delegate, "phase 3"-style, before
any "phase 2" target could be selected).

Bug: 202165481
Test: Manually verified unbundled-delegation still works.
Change-Id: Idfbd9022d3d195e7116b89c9bd2ddb207a15c0a3