Initial (inherited) tests for unbundled chooser.
This required disabling the #testEditImageLogs method in the base
class (ag/16520095) due to what could be a legitimate failure in the
unbundled implementation; we should investigate imminently.

Note this only brings the tests up to parity with the system
implementation - existing issues in the system ChooserActivityTest (e.g.
b/211669337) will presumably still occur in the unbundled version.

No test flakes are known to be specific to unbundling (among those that
are enabled after ag/16520095).

Bug: 202165481,211669337
Test: `atest UnbundledChooserActivityTest`
Change-Id: If8e2b896627661e59fa4a85aa3dd1eedb8eb91bc