Switch off sharing framework's "widgets"
This forks the ResolverDrawerLayout, which appears to be custom and
specific to our app. Any other dependencies from the framework package
com.android.internal.widget were old copies of androidx libraries,
and the "unbundled" IntentResolver is able to take dependencies on the
real versions instead. No dependencies into that framework package
remain after this CL.

These changes were excluded from ag/20001357 (as described in
go/chooser-fork-cl) in order to isolate any possible regressions (and
also to split up the review burden for the original fork CL).

There are no behavior changes *expected* from this CL, but any
edge-case discrepancies are hopefully improvements due to bugfixes or
other refinements that were put into the androidx libraries after the
framework copy was cut. The APIs themselves are supposed to be drop-in
compatible.

Test: atest IntentResolverUnitTests
Bug: 248566504
Change-Id: I8e02fd580d395446066fc2bb239a9bef833ef059