Unify APIs descending from ChooserTargetInfo.
Specifically, this pulls up the two new methods introduced by
SelectableTargetInfo (since we can easily specify a default behavior
for other kinds of ChooserTargetInfo) and adds @Override annotations
to show that all the remaining methods are now just overrides that
conform to the same overall interface. Finally, any clients that
"access" (but don't "create") instances of SelectableTargetInfo are
now able to perform all their same operations against the parent
ChooserTargetInfo, so all these call sites have been generalized to
remove the dependency on the narrower SelectableTargetInfo type.

As per go/chooser-targetinfo-cleanup this is one step in a broader
refactoring to simplify & ideally "flatten" this class hierarchy.

Test: atest IntentResolverUnitTests
Bug: 202167050
Change-Id: Id6c4f8e904d4390ed86cbc8c4966babd364b8209