Re-enable r8 optimization in IntentResolver
The previous "un-kept symbol" problem (with ChooserGridLayoutAdapter)
was inadvertently fixed by ag/20120577 (confirmed by going back to
check `apkdiffstats` before-and-after to note that the old crash is
due to a method that was only "un-kept" in the old version that still
shared framework RecyclerView). Manual review of other un-kept
symbols under `com.android.intentresolver.*` all seemed reasonable in
this new CL. For more info, see discussion on b/249159065.

Bug: 249159065, 249081140
Test: manual, and `atest IntentResolverUnitTests`
Change-Id: Ia6cf5dd80ba3ec16038225ccdac5203782c3db1f