Make "stacked" target membership immutable
This includes a readability improvement in `ChooserListAdapter` to
precompute the grouping for "consolidated" targets *before* building
them into a `MultiDisplayResolveInfo` so that we don't need to
accumulate targets into the respective stacks incrementally as we
process the list. This is the only time targets were being added into
a `MultiDisplayResolveInfo` -- they were already effectively final
after this one-time "consolidation" is complete.

Test: atest IntentResolverUnitTests
Bug: 202167050
Change-Id: I230131585a7087f7a839e58d1d5a2dbdf66c3712