Simplify ChooserActivityLogger.
This interface only had one implementation outside of tests, and the
test implementation was also unnecessarily complex. This CL
consolidates the "real" implementation into the base interface,
with injectable shims for the logging backend we'll use in tests.

After this CL, integration tests that want to verify logging sequences
can use a mock `ChooserActivityLogger` instead of a
`FakeChooserActivityLogger`, and then make assertions about the
(simpler) application-level events in that API. A new
`ChooserActivityLoggerTest` covers the translation of these events
into the appropriate representation for their `UiEventLogger` or
`FrameworkStatsLog` backend, so integration tests no longer need to
worry about that underlying detail.

In the past, the only integration tests that exercised logging had
been disabled in `ChooserActivityTest` due to flakes with
indeterminate logging sequences to be addressed in b/211669337.
Instead of attempting to patch the new (mock `ChooserActivityLogger`)
style into test code that's already broken and disabled, I just
deleted the relevant sections of those tests for now.

Test: atest IntentResolverUnitTests
Bug: 202167050, 211669337
Change-Id: Iaab551625284335469069bab8ee9a2d52fd955e6