Extract ChooserActivity.RoundedRectImageView
This is a trivial code move (diffs show only minor formatting
changes [EDIT: a later snapshot also gets rid of the `<view>` tags
in our layouts, which were only necessary to reference this as an inner
class, in favor of the more common practice of using an XML element
with the same fully-qualified name). As an inner class, it had already
been marked `static`, so there's no correctness concern (EDIT: on the
Java side...) The only "client" is the (soon-to-be-extracted) preview
UI, but instead of moving this to an inner class of the extracted
component, it seems reasonable for this class to stand alone -- it has
a simple design that's plausibly reusable in other applications, since
it doesn't really encapsulate any requirements that are specific to
Sharesheet.

Bug: 202167050
Test: atest IntentResolverUnitTests
Change-Id: I63833626925896a0baaa06d729233cd7037ac730