Extract ChooserActivity inner ViewHolder classes
These are all related to the inner ChooserGridAdapter, which I also
hope to lift out of the monolith eventually, but that's a more
complicated refactoring and this CL gets some low-hanging fruit out
of the way first.

Most of the classes were marked `static` or could have it trivially
added. I only had to (minorly) rework ItemViewHolder to inject the
callbacks; IMO this is a reasonable separation of the presentation
logic (in the view holder) from the application logic (injected from
the ChooserActivity).

Test: atest IntentResolverUnitTests
Bug: 202167050
Change-Id: Ic9707e079bb472dced270f54984e6e99bea5503a