[Chooser/ResolverActivity] Add option to show preview area even if there are no apps in the tab
Add a protected method that could be overriden
to change the default behavior that hides the
content preview area when there are not items
for the current user.

This is needed for the partial screen sharing
feature as we want to show the recents app selector
no matter if we have apps available for the
current profile or not. Recent app selector is
displayed in content preview area.

Test: atest com.android.intentresolver.UnbundledChooserActivityWorkProfileTest
Bug: 233348916
Change-Id: I0fa8da2dd4a690f5be1741d0936f776a3374aee7