Add optional text exclusion from media + text share
Add an option to the user to exclude text from being shared from a media
+ text sharing case.
The text is ecluded only if the primary send inten is used i.e. it is
possible for senders to specify alterntive intents that can also be
used.

Bug: 262277421
Test: manual testing of: (1) regualr image sharing, (2) image and
text/web-link sharing, (3) using of an alternative intent, (4) sharing
to a shortcut. All of those for one and two (+work) profiles.
Test: integration tests

Change-Id: Iebb33b951853c6588157fa92d3826b6a6e096591